requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
translations==0.1.3
urllib3==1.26.4
